Đã sửa lỗi và hoàn thành tính năng

+++ server/src/models/quatrinhhoc.js
@@ -24,7 +24,7 @@ module.exports = function(sequelize, DataTypes) {
       }
     },
     DiemTBHocKy: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT(4,2),
       allowNull: false
     }
   }, {
